Cleaned up query API, and added initial test template.
The :result-type argument of solr-query is deprecated.  It is
less useful if the result is pagenized, for the number of total
hits is lost.  Instead, the client can pass the result LXML to
utility functions such as solr-result->doc-nodes.

Alos added test template to make unit test run by
(asdf:oos 'asdf:test-op :solr).  The actual tests
needs to be written (currently it just fails).